+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/OAuth2AuthorizationServerAdapter.kt
@@ -2,10 +2,9 @@ package at.asitplus.wallet.lib.oidvci
 
 import at.asitplus.KmmResult
 import at.asitplus.openid.OAuth2AuthorizationServerMetadata
-import at.asitplus.openid.OidcUserInfoExtended
 import at.asitplus.wallet.lib.oauth2.RequestInfo
 import at.asitplus.wallet.lib.oauth2.TokenVerificationService
-import at.asitplus.wallet.lib.openid.AuthenticationResponseResult
+import kotlinx.serialization.json.JsonObject
 
 /**
  * Used in OID4VCI by [CredentialIssuer] to obtain user data when issuing credentials using OID4VCI.
@@ -37,7 +36,7 @@ interface OAuth2AuthorizationServerAdapter {
         credentialIdentifier: String?,
         credentialConfigurationId: String?,
         request: RequestInfo?,
-    ): KmmResult<OidcUserInfoExtended>
+    ): KmmResult<JsonObject>
 
 }